requirements.txt
spotipy
python-dotenv
request
